move unsused files to deprecated
