+++ deprecated/main/common/Carousal/SmallCarousel.tsx
@@ -1,6 +1,6 @@
 import styles from "./SmallCarousel.module.scss";
 import classNames from "classnames/bind";
-import { IEventSummary } from "../../../../types/event";
+import { IEventSummary } from "../../../types/event";
 import EventThumbnail from "../EventThumbnail/EventThumbnail";
 
 interface Props {
